+++ Classy.Models/PurchaseOption.cs
@@ -10,12 +10,12 @@ namespace Classy.Models
         public string Content { get; set; }
         public Dictionary<string, string> VariantProperties { get; set; } // Key: Size, Color, Model, etc. Value: Smal, Medium, Large, etc.
         public string SKU { get; set; }
-        public double Price { get; set; }
-        public double Quantity { get; set; }
-        public double? CompareAtPrice { get; set; }
+        public decimal Price { get; set; }
+        public int Quantity { get; set; }
+        public decimal? CompareAtPrice { get; set; }
         public MediaFile[] MediaFiles { get; set; }
         public string ProductUrl { get; set; }
-        public double NeutralPrice { get; set; }
+        public decimal NeutralPrice { get; set; }
         public string Height { get; set; }
         public string Width { get; set; }
         public string Depth { get; set; }
