+++ Classy.Repository/ITaxCalculator.cs
@@ -1,17 +1,9 @@
 ﻿using Classy.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Classy.Repository
 {
     public interface ITaxCalculator
     {
-        double CalculateTax(
-            Profile merchantProfile, 
-            double amount, 
-            PhysicalAddress address);
+        decimal CalculateTax(Profile merchantProfile, decimal amount, PhysicalAddress address);
     }
 }
