+++ classifieds/Extentions/PricingInfoViewExtensions.cs
@@ -1,16 +1,14 @@
 ﻿using Classy.Models;
 using Classy.Models.Response;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Web;
 using ServiceStack.Common;
 
 namespace classy
 {
     public static class PricingInfoViewExtentions
     {
-        public static PricingInfoView ToPricingInfoView(this PricingInfo from, double adjustRate)
+        public static PricingInfoView ToPricingInfoView(this PricingInfo from, decimal adjustRate)
         {
             var to = from.TranslateTo<PricingInfoView>();
             to.BaseOption = from.BaseOption.TranslateTo<PurchaseOptionView>();
